requirements.txt
supabase==2.3.4
pandas==2.1.4
numpy>=1.26.0
python-dotenv==1.0.0
scikit-learn>=1.3.2
joblib>=1.3.2
Flask==3.1.2
flask-cors==6.0.1
pypdf==5.4.0
Pillow==11.2.1
pytesseract==0.3.13
